라우터 정리 및 fetch 1차 수정
라우터 정리 및 fetch 1차 수정
