+++ src/pages/mypage/MyPlabPage.jsx
@@ -4,8 +4,11 @@ import MatchCompleted from "../../components/mypage/myplab/MatchCompleted";
 import CalendarModal from "../../components/mypage/myplab/CalendarModal";
 import TabNavigation from "../../components/mypage/myplab/TabNavigation";
 import "./page-style.css";
+import { config } from '../../config';
+
 
 const MyPlabPage = () => {
+  const api = config.aws.ec2_host_user
   const [selectedDate, setSelectedDate] = useState(null); // 선택된 날짜
   const [selectedMonth, setSelectedMonth] = useState(new Date()); // 선택된 월
   const [activeTab, setActiveTab] = useState("schedule"); // 현재 활성화된 탭 ('schedule' or 'completed')
@@ -27,7 +30,7 @@ const MyPlabPage = () => {
   useEffect(() => {
     const fetchScheduleData = async () => {
       try {
-        const response = await fetch("http://127.0.0.1:8080/mypage/myplab", {
+        const response = await fetch(`${api}/mypage/myplab`, {
           method: "POST",
           headers: {
             "Content-Type": "application/json",
@@ -41,15 +44,15 @@ const MyPlabPage = () => {
         }
 
         const data = await response.json();
-        setScheduleData(data); // JSON 구조를 상태에 저장
-        console.log("Schedule Data:", data);
+        setScheduleData(data);
+        // console.log("Schedule Data:", data);
       } catch (error) {
         console.error("Error fetching match schedule:", error.message);
       }
     };
 
     fetchScheduleData();
-  }, [userId, token]);
+  }, [userId, token, api]);
 
   const handleDateSelect = (date) => {
     setSelectedDate(date); // 선택된 날짜를 상태에 저장
