+++ src/router/About.jsx
@@ -1,13 +1,8 @@
 import React from "react";
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import AboutPage from "../pages/root/About";
 
 function About() {
-  return (
-    <Routes>
-      <Route path="/" element={<AboutPage />} />
-    </Routes>
-  );
+  return <AboutPage /> ;
 }
 
 export default About;
