+++ src/router/Cs.jsx
@@ -1,13 +1,8 @@
 import React from "react";
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import CSpage from "../pages/root/CSpage";
 
 function Cs() {
-  return (
-    <Routes>
-      <Route path="/" element={<CSpage />} />
-    </Routes>
-  );
+  return <CSpage />;
 }
 
 export default Cs;
