+++ src/router/Main.jsx
@@ -1,13 +1,8 @@
 import React from 'react';
-import { Routes, Route } from 'react-router-dom';
 import MainPage from '../pages/main/MainPage';
 
 const Main = () => {
-    return (
-      <Routes>
-        <Route path="/" element={<MainPage />} />
-      </Routes>
-    );
-  };
+  return <MainPage />;
+};
 
 export default Main;
