+++ src/router/NotFound.jsx
@@ -3,7 +3,7 @@ import React from "react";
 function NotFound() {
   return (
     <>
-      <h1>404 요청 분기 처리</h1>
+      <h1>404 not found</h1>
     </>
   );
 }
