+++ src/router/Notice.jsx
@@ -1,13 +1,8 @@
 import React from "react";
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import NoticePage from "../pages/root/Noticepage";
 
 function Notice() {
-  return (
-    <Routes>
-      <Route path="/" element={<NoticePage />} />
-    </Routes>
-  );
+  return <NoticePage />;
 }
 
 export default Notice;
